Add support for composite unique constraints

+++ piccolo/apps/migrations/auto/schema_differ.py
@@ -21,6 +21,7 @@ from piccolo.apps.migrations.auto.serialisation import (
     UniqueGlobalNames,
     serialise_params,
 )
+from piccolo.columns.constraints import UniqueConstraint
 from piccolo.utils.printing import get_fixed_length_string
 
 
@@ -270,8 +271,12 @@ class SchemaDiffer:
             used_drop_column_names: t.List[str] = []
 
             for add_column in delta.add_columns:
+                if add_column.column_class == UniqueConstraint:
+                    continue
                 for drop_column in delta.drop_columns:
                     if drop_column.column_name in used_drop_column_names:
+                        continue
+                    if drop_column.column_class == UniqueConstraint:
                         continue
 
                     user_response = self.auto_input or input(
@@ -508,6 +513,11 @@ class SchemaDiffer:
                     )
 
                 if alter_column.old_column_class is not None:
+                    if alter_column.old_column_class == UniqueConstraint:
+                        print(
+                            f"You cannot ALTER `{alter_column.column_name}` unique constraint! At first, delete it, then create the new one."  # noqa: E501
+                        )
+                        continue
                     extra_imports.append(
                         Import(
                             module=alter_column.old_column_class.__module__,
@@ -538,6 +548,7 @@ class SchemaDiffer:
     @property
     def drop_columns(self) -> AlterStatements:
         response = []
+        extra_imports: t.List[Import] = []
         for table in self.schema:
             snapshot_table = self._get_snapshot_table(table.class_name)
             if snapshot_table:
@@ -552,14 +563,29 @@ class SchemaDiffer:
                 ):
                     continue
 
+                column_class = column.column_class
+                extra_imports.append(
+                    Import(
+                        module=column_class.__module__,
+                        target=column_class.__name__,
+                        expect_conflict_with_global_name=getattr(
+                            UniqueGlobalNames,
+                            f"COLUMN_{column_class.__name__.upper()}",  # noqa: E501
+                            None,
+                        ),
+                    )
+                )
+
                 schema_str = (
                     "None" if column.schema is None else f'"{column.schema}"'
                 )
 
                 response.append(
-                    f"manager.drop_column(table_class_name='{table.class_name}', tablename='{table.tablename}', column_name='{column.column_name}', db_column_name='{column.db_column_name}', schema={schema_str})"  # noqa: E501
+                    f"manager.drop_column(table_class_name='{table.class_name}', tablename='{table.tablename}', column_name='{column.column_name}', db_column_name='{column.db_column_name}', schema={schema_str}, column_class={column_class.__name__})"  # noqa: E501
                 )
-        return AlterStatements(statements=response)
+        return AlterStatements(
+            statements=response, extra_imports=extra_imports
+        )
 
     @property
     def add_columns(self) -> AlterStatements:
